feat(connector): multi-platform connector architecture — shared core + REST API + OpenAI wiring
Extract all tool handlers and auth logic from mcp-server into _shared/connector-core.ts (shared module).
MCP server becomes a thin JSON-RPC protocol wrapper importing from shared core.

Add connector-api edge function — REST endpoints (OpenAPI 3.1) consumed by OpenAI GPT Actions
and Google Gemini Extensions. Same auth, same privacy tiers, same enrichment pipeline.

Rename mcp_tokens/mcp_auth_codes/mcp_clients → connector_tokens/connector_auth_codes/connector_clients.
Add platform column for multi-platform analytics. Add per-token rate limiting (60 req/min).

Add OpenAI redirect URIs to OAuth server. Platform auto-detected from redirect_uri.

Version bumps: mcp-server 0.3.0→0.4.0, mcp-oauth 0.1.1→0.2.0, connector-api 0.1.0 (new).
